agentic_ke/case01_animals-onto-agents/requirements.txt
anthropic>=0.25.0
owlready2>=0.46
python-dotenv>=1.0.0
pydantic>=2.0.0
